✨ Generate vuejs view and style file

+++ generators/app/index.js
@@ -1,21 +1,49 @@
 'use strict';
 const Generator = require('yeoman-generator');
 const chalk = require('chalk');
-const yosay = require('yosay');
 
 module.exports = class extends Generator {
-  prompting() {
-    // Have Yeoman greet the user.
-    this.log(
-      yosay(`Welcome to the fantabulous ${chalk.red('generator-creamio-vuejs')} generator!`)
-    );
 
+  constructor(args, opts) {
+    // Calling the super constructor is important so our generator is correctly set up
+    super(args, opts);
+
+    this.views = this.fs.
+  }
+  prompting() {
+    this.log(`\nWelcome to the ${chalk.red('CreamIO Vue.js')} generator !\n`);
+    
     const prompts = [
       {
+        type: 'list',
+        name: 'objectType',
+        message: 'What do you want to generate ?',
+        choices: ['View', 'Component'],
+        default: 0
+      },
+      {
+        type: 'input',
+        name: 'objectName',
+        message: 'What is the name of your element ?'
+      },
+      // VIEW PART
+      {
         type: 'confirm',
-        name: 'someAnswer',
-        message: 'Would you like to enable this option?',
-        default: true
+        name: 'createStyle',
+        message: 'Should we create a SCSS file for this view ?',
+        default: true,
+        when: hash => hash.objectType === 'View'
+      },
+      // COMPONENT PART
+      {
+        type: 'confirm',
+        name: 'createStyle',
+        message: 'Should we create a SCSS file for this component ?',
+        default: true,
+        when: hash => hash.objectType === 'Component'
+      },
+      {
+        type:
       }
     ];
 
@@ -26,13 +54,38 @@ module.exports = class extends Generator {
   }
 
   writing() {
-    this.fs.copy(
-      this.templatePath('dummyfile.txt'),
-      this.destinationPath('dummyfile.txt')
-    );
-  }
-
-  install() {
-    this.installDependencies();
+    /*
+     * VIEW PART
+     */
+    if (this.props.objectType === 'View') {
+      if (this.fs.exists(this.destinationPath('src/views/' + this.props.objectName + '.vue'))) {
+        this.env.error('The following file : ' + chalk.red(this.destinationPath('src/views/' + this.props.objectName + '.vue')) + ' already exists.');
+        return;
+      }
+      this.fs.copyTpl(
+        this.templatePath('views/_view.vue'),
+        this.destinationPath('src/views/' + this.props.objectName + '.vue'),
+        {
+          name: this.props.objectName,
+          createStyle: this.props.createStyle
+        }
+      );
+      // VIEW STYLE
+      if (this.props.createStyle === true) {
+        if (this.fs.exists(this.destinationPath('src/scss/views/' + this.props.objectName + '.scss'))) {
+          this.env.error('The following file : ' + chalk.red(this.destinationPath('src/scss/views/' + this.props.objectName + '.scss')) + ' already exists.');
+          return;
+        }
+        this.fs.copy(
+          this.templatePath('views/_style.scss'),
+          this.destinationPath('src/scss/views/' + this.props.objectName + '.scss')
+        );
+      }
+    /*
+     * COMPONENT PART
+     */
+    } else if (this.props.objectType === 'Component') {
+      this.log('issou');
+    }
   }
 };
